Photon: Use correct dims for older images.
If you are outputting a post thumbnail at a custom size that did not exist when the image was uploaded (thus does not exist per WP), Photon in 3.8.1 would output it at full-size with the full-size values as the height/width on the img tag.

This PR uses a core function to confirm we only return a reasonable size.