Photon: Selectively crop.
Prior to 3.8.1, the `crop` arg was determined from the `self::image_sizes()` function, except for `full` size images, which was declared in this section of code.

Full size images render the same with `fit` or `resize` by definition, making this declaration not needed.

In 3.8.1, we always set `image_args['crop'] = true;` resulting in cases when an intermediate size was not found, we would use the full-size image and crop it the specific size indicated, instead of relying on the crop arg that was set when the image size was originally declared in WordPress.

Example: If there was a `example_size` image size, set to 700 x 100, with crop false, the expected return would be `fit` in 3.8, e.g.
https://i1.wp.com/kraft.im/uploads/2014/08/snugharbor.jpeg?fit=700%2C100

In 3.8.1, we would return: https://i1.wp.com/kraft.im/uploads/2014/08/snugharbor.jpeg?resize=700%2C100